refactor: validate 역할 위임
